psalm level 5 fixes (#34)
* psalm level 5 fixes

* Adding null as allowed param type for where queries

It's possible in Doctrine to pass `null` as a param type in a where query (just the simple queries) and Doctrine will convert that to the proper SQL syntax for the underlying RDMS.

This doesn't work for the whereIn methods, as `null` would result in an empty array when it's casted, which Doctrine checks for and throws an exception to prevent "WHERE IN()" queries.

The `$params !== null` checks aren't necessary here as the `is_array` check won't pass if the value is null.

+++ lib/Doctrine/Cli/AnsiColorFormatter.php
@@ -154,7 +154,7 @@ class Doctrine_Cli_AnsiColorFormatter extends Doctrine_Cli_Formatter
             return $text;
         }
 
-        $subsize = floor(($size - 3) / 2);
+        $subsize = (int) floor(($size - 3) / 2);
 
         return substr($text, 0, $subsize) . $this->format('...', 'INFO').substr($text, -$subsize);
     }
